pyproject.toml
[project]
name = "naver-search-mcp"
version = "0.1.0"
description = "Simple Naver Search MCP server using Naver Open API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
naver-search-mcp = "naver_search.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/naver_search"]
